working k8s sd
